requirements.txt
celery[rabbitmq]
boto3
minio
snowflake-connector-python
beautifulsoup4
requests
pika
pytest
flake8
pydantic
python-dotenv
playwright
fake-useragent
praw
datetime
pyyaml
autoflake
black
